chore: move code around
This relocates the `src/boostrap/create_bm25.rs` source to
`src/api/admin.rs`, which I believe is a more sensibile location and
name.

Additionally it moves the `#[pg_extern] schema()` function from
`src/api/builder_fns.rs` to this new(ly renamed) `admin.rs`, as it's an
administrative function, not a query builder function.
