+++ pg_search/src/bootstrap/mod.rs
@@ -15,5 +15,4 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-mod create_bm25;
 mod test_table;
